added a basic version of the reporter
Work needs to be done on adding this to a significant sized test suite
and seeing if the colour changes need to be made on test completion or
on suite completion.

At the moment it is using the first LED that it finds on the stick but
that will be changed to make use of all available LED on a stick. This
will be overkill on the bigger blinksticks but great on the smaller
ones.